fix(sdk): rename mcp service module to avoid conflicts
